+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - removed walruses (support for python 3.7)
 - fixed bugs in various converter methods
 - fix for silent setup when base url provided
+- deals with default_revision all false
 
 ## [0.1.5]