+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.1.6]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.2.0]
 
 ### Added