+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 
-## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.2.0]
+## [Latest](https://github.com/int-brain-lab/ONE/commits/main) [0.3.0]
+
+### Modified
+
+- webclient encodes characters in URL query parameters
+- docstrings added to ONE factory and classes
+
+## [0.2.0]
 
 ### Added